Chasing pandas 2.0 issues
Looks like issues are less related to the use of ufunc (e.g. np.log10) on Object type series
(which has be deprecated since before Pandas 1.5) but rather related to assignments with
`df.loc[:,col] = RHS` casting RHS into the dtype of `df[col]` and, therefore,
when RHS is expected to update the series dtype (e.g. to float or int [using apply()
or astype()]) the action does nothing. Conversion of pst data cols was not occuring as expected
and subsequent ufunc calls (e.g. log10) were failing.
A big WTF but seems to work if the assignment is of the form `df[col]=RHS`.

Have stripped back most of JTW mods after making this change in pst_handler.py